+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ This is the offical release of vpinfe v2.0.3
 ### What's New
 - More managerUI clearing/seperation.  Logs now on its own page.  added vpx logs
 - QRCode in the mainmenu for linking to remote.
+- VPinPlay Player QR code.
 
 ### Fixes
 None